Move common behaviour code to a seperate module
