+++ lib/membrane/source.ex
@@ -19,7 +19,7 @@ defmodule Membrane.Source do
 
   defmacro __using__(_) do
     quote location: :keep do
-      use Membrane.Element
+      use Membrane.Element.Base
       use Membrane.Element.WithOutputPads
 
       @impl true
